+++ tests/fixtures/issue-1712-angular-for-class-member/src/app.component.ts
@@ -1,0 +1,20 @@
+import { Component } from '@angular/core'
+import { Util } from './utils/Util'
+
+@Component({
+  selector: 'app-root',
+  standalone: true,
+  template: `
+    <ul>
+      @for (util of utils; track util) {
+        <li>{{ util.getName() }} {{ util.getter }} {{ util.property }}</li>
+      }
+    </ul>
+    <ul>
+      <li *ngFor="let util of utils">{{ util.getName() }}</li>
+    </ul>
+  `,
+})
+export class AppComponent {
+  utils: Util[] = [new Util()]
+}